QA review fixes for #91
- automations: wrap scheduler.stop() in try/finally so the in-process MCP
  transport always closes, even if scheduler ever grows a code path that
  throws. Today the call is just `clearInterval`; the guard removes the
  asymmetry between "scheduler error" and "leaked transport."
- Update doc comments that still mentioned `InlineSource` after the
  migration: `bundles/types.ts`, `bundles/defaults.ts`,
  `tools/platform/files.ts`, `tools/platform/settings.ts`,
  `tools/platform/index.ts`, `tools/platform-resources/conversations/
  browser.ts`, `tools/validate-input.ts`, `tools/in-process-app.ts`,
  `bundles/automations/src/executor.ts`, `bundles/automations/src/
  schemas.ts`. Zero `InlineSource` mentions remain in `src/`.
- Extract `bytesToBase64` to `src/util/base64.ts` so both the API resource
  envelope path and the in-process MCP server resource handler share one
  implementation. Avoids the upward `tools/ → api/` dependency that
  re-exporting from `handlers.ts` would require.
- Delete `getCoreResource()` — after the C migration its only consumer was
  `test/unit/core-resources.test.ts`. Test now asserts via
  `buildCoreResourceMap()` and the `ui://nb/<path>` URI keys, which is the
  same path production uses.
