+++ src/bundles/types.ts
@@ -188,8 +188,8 @@ export interface BundleInstance {
   /**
    * Workspace that owns this instance. Required — every bundle instance
    * belongs to exactly one workspace. Global/platform sources are
-   * represented as InlineSource, not BundleInstance, so they never reach
-   * this type.
+   * in-process MCP servers (`defineInProcessApp`), not BundleInstance,
+   * so they never reach this type.
    */
   wsId: string;
   /** Absolute path to the entity data root (e.g., {wsDir}/data/{bundle}/apps/crm/data). Resolved at startup. */
