+++ src/tools/core-resources/index.ts
@@ -42,15 +42,6 @@ const resources: Record<string, () => string> = {
 };
 
 /**
- * Get an HTML resource by path name.
- * Returns the HTML string or null if the resource does not exist.
- */
-export function getCoreResource(path: string): string | null {
-  const factory = resources[path];
-  return factory ? factory() : null;
-}
-
-/**
  * Build a Map of all core resources for use with the `nb` in-process MCP
  * source. Keys are full `ui://nb/<path>` URIs — the form servers and
  * clients both use over the protocol — so lookups via
