+++ src/tools/platform-resources/conversations/browser.ts
@@ -1,7 +1,7 @@
 /**
  * Re-export the conversation browser HTML from the bundle source.
  *
- * This allows the platform InlineSource to serve the same UI
- * without duplicating the HTML.
+ * This allows the platform `conversations` in-process MCP source to
+ * serve the same UI without duplicating the HTML.
  */
 export { BROWSER_HTML } from "../../../bundles/conversations/src/ui/browser.ts";
